+++ src/dummynet/host_shell.py
@@ -50,7 +50,7 @@ class HostShell(object):
         stdout, stderr = process.communicate()
         returncode = process.wait()
 
-        result = run_info.RunInfo(
+        info = run_info.RunInfo(
             cmd=cmd,
             cwd=cwd,
             stdout=stdout,
@@ -60,10 +60,10 @@ class HostShell(object):
             is_daemon=False,
         )
 
-        if result.returncode != 0:
-            raise errors.RunInfoError(result=result)
+        if info.returncode != 0:
+            raise errors.RunInfoError(info=info)
 
-        return result
+        return info
 
     def run_async(self, cmd: str, daemon=False, cwd=None):
         """Run an asynchronous command (non-blocking).
@@ -89,7 +89,7 @@ class HostShell(object):
             text=True,
         )
 
-        result = run_info.RunInfo(
+        info = run_info.RunInfo(
             cmd=cmd,
             cwd=cwd,
             stdout="",
@@ -99,6 +99,6 @@ class HostShell(object):
             is_daemon=daemon,
         )
 
-        self.process_monitor.add_process(popen, result)
+        self.process_monitor.add_process(popen, info)
 
-        return result
+        return info